attendance_system/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
numpy==1.26.4
opencv-contrib-python==4.8.1.78
face-recognition==1.3.0
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
requests==2.31.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0